# Make snapshot load faster