--- a/src/Service/KeeperStore.h
+++ b/src/Service/KeeperStore.h
@@ -84,7 +84,7 @@ struct KeeperNodeWithPath
 /// Map for data tree, it is thread safe with read write lock.
 /// ConcurrentMap is a two-level unordered_map which is designed
 /// to reduce latency for unordered_map scales.
-template <typename Element, unsigned NumBlocks>
+template <typename Element, unsigned NumBuckets>
 class ConcurrentMap
 {
 public:
@@ -149,7 +149,7 @@ public:
     };
 
 private:
-    std::array<InnerMap, NumBlocks> maps_;
+    std::array<InnerMap, NumBuckets> maps_;
     std::hash<String> hash_;
 
 public:
@@ -161,9 +161,9 @@ public:
     size_t count(const String & key) { return get(key) != nullptr ? 1 : 0; }
     bool erase(String const & key) { return mapFor(key).erase(key); }
 
-    InnerMap & mapFor(String const & key) { return maps_[hash_(key) % NumBlocks]; }
-    UInt32 getBlockIndex(String const & key) { return hash_(key) % NumBlocks; }
-    UInt32 getBlockNum() const { return NumBlocks; }
+    InnerMap & mapFor(String const & key) { return maps_[hash_(key) % NumBuckets]; }
+    UInt32 getBucketIndex(String const & key) { return hash_(key) % NumBuckets; }
+    UInt32 getBucketNum() const { return NumBuckets; }
     InnerMap & getMap(const UInt32 & index) { return maps_[index]; }
 
     void clear()
@@ -185,9 +185,9 @@ public:
 class KeeperStore
 {
 public:
-    /// block num for ConcurrentMap
-    static constexpr int MAP_BLOCK_NUM = 16;
-    using Container = ConcurrentMap<KeeperNode, MAP_BLOCK_NUM>;
+    /// bucket num for ConcurrentMap
+    static constexpr int MAP_BUCKET_NUM = 16;
+    using Container = ConcurrentMap<KeeperNode, MAP_BUCKET_NUM>;
 
     using ResponsesForSessions = std::vector<ResponseForSession>;
     using KeeperResponsesQueue = ThreadSafeQueue<ResponseForSession>;
@@ -203,7 +203,10 @@ public:
     using SessionIDs = std::vector<int64_t>;
 
     using Watches = std::unordered_map<String /* path, relative of root_path */, SessionIDs>;
-    using BlocksEdges = std::array<Edges, MAP_BLOCK_NUM>;
+
+    /// Hold Edges in different Buckets based on the parent node's bucket number.
+    /// It should be used when load snapshot to built node's childrenSet in parallel without lock.
+    using BucketEdges = std::array<Edges, MAP_BUCKET_NUM>;
 
     /// global session id counter, used to allocate new session id.
     /// It should be same across all nodes.
@@ -301,8 +304,8 @@ public:
     /// Build path children after load data from snapshot
     void buildPathChildren(bool from_zk_snapshot = false);
 
-    // Build childrenset for the node in specified block after load data from snapshot.
-    void buildBlockChildren(const std::vector<BlocksEdges> & all_objects_edges, UInt32 block_idx);
+    // Build childrenSet for the node in specified bucket after load data from snapshot.
+    void buildBucketChildren(const std::vector<BucketEdges> & all_objects_edges, UInt32 bucket_idx);
 
     void finalize();
 
@@ -342,7 +345,7 @@ public:
     uint64_t getApproximateDataSize() const
     {
         UInt64 node_count = container.size();
-        UInt64 size_bytes = container.getBlockNum() * sizeof(Container::InnerMap) /* Inner map size */
+        UInt64 size_bytes = container.getBucketNum() * sizeof(Container::InnerMap) /* Inner map size */
             + node_count * 8 / 0.75 /*hash map array size*/
             + node_count * sizeof(KeeperNode) /*node size*/
             + node_count * 100; /*path and child of node size*/