--- a/src/Service/NuRaftLogSnapshot.h
+++ b/src/Service/NuRaftLogSnapshot.h
@@ -104,16 +104,16 @@ private:
     void getObjectPath(ulong object_id, String & path);
 
     /// Parse an snapshot object. We should take the version from snapshot in general.
-    void parseObject(KeeperStore & store, String obj_path);
+    void parseObject(KeeperStore & store, String obj_path, BlocksEdges &);
 
     /// Parse batch header in an object
     /// TODO use internal buffer
     void parseBatchHeader(ptr<std::fstream> fs, SnapshotBatchHeader & head);
 
     /// Parse a batch /// TODO delete
-    void parseBatchBody(KeeperStore & store, char * batch_buf, size_t length, SnapshotVersion version_);
+    void parseBatchBody(KeeperStore & store, const String &, BlocksEdges &, SnapshotVersion version_);
     /// Parse a batch
-    void parseBatchBodyV2(KeeperStore & store, char * buf, size_t length, SnapshotVersion version_);
+    void parseBatchBodyV2(KeeperStore & store, const String &, BlocksEdges &, SnapshotVersion version_);
 
     /// Serialize whole data tree /// TODO delete
     size_t serializeDataTree(KeeperStore & storage);
@@ -167,6 +167,8 @@ private:
     UInt64 last_log_term;
 
     std::map<ulong, String> objects_path;
+
+    std::vector<BlocksEdges> all_objects_edges;
 
     /// Appended to snapshot file name
     String curr_time;