… .github/workflows/build-and-test-macos.yml
… src/Service/SnapshotCommon.h
… src/Service/tests/gtest_raft_snapshot.cpp