### Which issues of this PR fixes:
<!-- Usage: `Fixes #<issue number>` -->
This PR try to fix #221
For nodes count 58680021, snapshot size 12GB

|                     | Before(sec) | After(sec) |
| ------------------- | ----------- | ---------- |
| Parse from disk     | 90          | 70         |
| Build children path | 90          | 9          |
### Change log:
<!-- (Please describe the changes you have made in details. -->

Optimizations:

1. Reduced one buffer copy and destruction.
2. Storage edges in bucketed based on the parent node's bucket number, and built childrenSet in parallel.
3. Reserved space for childrenSet in advance.