@JackyWoo 2024-03-26
@lzydmxy Nice optimization! Please describe your solution.