@JackyWoo 2024-03-26
Is there any optimization for parsing data from disk?