@lzydmxy 2024-03-26
> @lzydmxy Nice optimization! Please describe your solution.
Here are three optimization points:
Reduced one buffer copy and destruction.
Storage edges in bucketed based on the parent node's bucket number, and built childrenSet in parallel.
Reserved space for childrenSet in advance.