@JackyWoo 2024-03-28
We should not use Apple clang. Let us update the CI env.